Restructure code and make it pip installable
Currently the code structure was not following best practices (eg. not
using console_scripts entry_points, not having a main() method which
can be executed in the top-level code environment[0]).

This commit restructures the code:
- move the code from uefivars.py to pyuefivars/__init__.py and add
  there some functions (_parser(), main()) to have a cleaner structure
- add an console_scripts entry_point to setup.py . That entry_point is
  named uefivars.py to match the AWS documentation[1] when users
  install via pip and want to follow docs (the difference would still
  be calling "uefivars.py" instead of "./uefivars.py" but needs to
  change in the docs)
- use the pyuefivars/__init__.py:main function in uefivars.py so users
  can still use "./uefivars.py" to run the program.

[0]
https://docs.python.org/3/library/__main__.html#what-is-the-top-level-code-environment
[1]
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/uefi-secure-boot-optionB.html
