+++ setup.py
@@ -9,7 +9,7 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="uefivars",
+    name="pyuefivars",
     version="0.1",
     author="Amazon Web Services",
     author_email="graf@amazon.com",
@@ -18,6 +18,11 @@ setuptools.setup(
     url="https://github.com/awslabs/python-uefivars",
     packages=setuptools.find_packages(),
     install_requires=[ 'crc32c' ],
+    entry_points={
+        'console_scripts': [
+            'uefivars.py = pyuefivars:main',
+        ],
+    },
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
